create page components
