+++ config/storyblok.ts
@@ -4,11 +4,16 @@ import StoryblokAudio from '../components/storyblok/StoryblokAudio';
 import StoryblokButton from '../components/storyblok/StoryblokButton';
 import StoryblokCard from '../components/storyblok/StoryblokCard';
 import StoryblokCarousel from '../components/storyblok/StoryblokCarousel';
+import StoryblokCoursePage from '../components/storyblok/StoryblokCoursePage';
 import StoryblokFaqs from '../components/storyblok/StoryblokFaqs';
 import StoryblokImage from '../components/storyblok/StoryblokImage';
+import StoryblokPage from '../components/storyblok/StoryblokPage';
+import StoryblokPageSection from '../components/storyblok/StoryblokPageSection';
 import StoryblokQuote from '../components/storyblok/StoryblokQuote';
 import StoryblokRow from '../components/storyblok/StoryblokRow';
 import StoryblokRowColumnBlock from '../components/storyblok/StoryblokRowColumnBlock';
+import StoryblokSessionIbaPage from '../components/storyblok/StoryblokSessionIBAPage';
+import StoryblokSessionPage from '../components/storyblok/StoryblokSessionPage';
 import StoryblokSpacer from '../components/storyblok/StoryblokSpacer';
 import StoryblokStatement from '../components/storyblok/StoryblokStatement';
 import StoryblokVideo from '../components/storyblok/StoryblokVideo';
@@ -30,5 +35,10 @@ export const storyblok = storyblokInit({
     accordion: StoryblokAccordion,
     carousel: StoryblokCarousel,
     spacer: StoryblokSpacer,
+    page_section: StoryblokPageSection,
+    page: StoryblokPage,
+    course: StoryblokCoursePage,
+    session: StoryblokSessionPage,
+    session_iba: StoryblokSessionIbaPage,
   },
 });
